Scripts/CoilfangReservoir: Changes here and there
